Add and adjust some comments for review
