+++ src/cli/commands/search.rs
@@ -13,8 +13,8 @@ use crate::{
             deprecation,
             logging::error,
             not_found,
-            standard_print::{self, MapStyled},
-            styled::Styled,
+            standard_print::{self},
+            styled::{MapStyled, Styled},
         },
     },
     config::Config,
