+++ src/cli/commands/update.rs
@@ -10,8 +10,8 @@ use crate::{
             QuestionResponse, ask_user, grid,
             logging::error,
             not_found,
-            standard_print::{self, MapStyled},
-            styled::Styled,
+            standard_print::{self},
+            styled::{MapStyled, Styled},
         },
         parameter_checks,
     },
